requirements.txt
pandas
numpy
seaborn
matplotlib
tensorflow
nltk
scikit-learn
Flask
ensure
python-box
fastapi
#-e .